requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
audioop-lts==0.2.1
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
durationpy==0.9
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
# grpcio==1.67.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-text-splitters==0.3.4
langsmith==0.2.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
moviepy==2.1.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.19.0
numpy==2.2.0
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
posthog==3.7.4
proglog==0.1.10
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.12.0
SQLAlchemy==2.0.36
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
wrapt==1.17.0
yarl==1.18.3
youtube-transcript-api==0.6.3
yt-dlp==2024.12.13
zipp==3.21.0
